added new map component
