+++ app.json
@@ -17,7 +17,12 @@
       "googleServicesFile": "./GoogleService-Info.plist",
       "bundleIdentifier": "com.kepleris.rescue",
       "usesAppleSignIn": true,
-      "iosMapsApiKey": "AIzaSyDkO9yZTw1mmEBrc5Vy6mIO1e5gAUfbG94"
+      "iosMapsApiKey": "AIzaSyDkO9yZTw1mmEBrc5Vy6mIO1e5gAUfbG94",
+      "NSLocationWhenInUseUsageDescription": "Esta aplicación necesita acceder a tu ubicación para mostrarte el mapa.",
+      "NSLocationAlwaysUsageDescription": "Esta aplicación necesita acceder a tu ubicación en segundo plano.",
+      "infoPlist": {
+        "UIBackgroundModes": ["location", "fetch", "remote-notification"]
+      }
     },
     "android": {
       "adaptiveIcon": {
@@ -27,7 +32,8 @@
         "androidMapsApiKey": "AIzaSyDkO9yZTw1mmEBrc5Vy6mIO1e5gAUfbG94"
       },
       "googleServicesFile": "./google-services.json",
-      "package": "com.kepleris.rescue"
+      "package": "com.kepleris.rescue",
+      "permissions": ["android.permission.SCHEDULE_EXACT_ALARM"]
     },
     "web": {
       "bundler": "metro",
@@ -39,6 +45,7 @@
       "@react-native-google-signin/google-signin",
       "expo-apple-authentication",
       "expo-secure-store",
+      "expo-localization",
       [
         "expo-location",
         {
