+++ app/(screens)/index.tsx
@@ -5,36 +5,11 @@ import { Button, FlatList, View, Text } from 'react-native';
 import { CategoryList } from '../../components/CategoryList';
 import { SearchBar } from '../../components/SearchBar';
 import { userLocationContext } from '../../src/context/userLocationContext';
-import * as Location from 'expo-location';
 import { useState, useEffect } from 'react';
-import  GoogleMapScreen from './googleMapScreen';
 
 export default function homeScreen() {
   const router = useRouter();
   const { signOut } = useSession();
-
-  const [location, setLocation] = useState<Location.LocationObject | null>(null);
-  const [errorMsg, setErrorMsg] = useState<string | null>(null);
-  useEffect(() => {
-    (async () => {
-
-      let { status } = await Location.requestForegroundPermissionsAsync();
-      if (status !== 'granted') {
-        setErrorMsg("Permission to access location was denied");
-        return;
-      }
-
-      let location = await Location.getCurrentPositionAsync({});
-      setLocation(location);
-    })();
-  }, []);
-
-  let text = 'Waiting..';
-  if (errorMsg) {
-    text = errorMsg;
-  } else if (location) {
-    text = JSON.stringify(location);
-  }
 
 
   const sampleProduct: ProductType = {
@@ -61,12 +36,9 @@ export default function homeScreen() {
       />
       <View>
         <Button title="Go to company profile" onPress={() => router.push('./companyScreen')} />
+        <Button title="Go to google map" onPress={() => router.push('./googleMapScreen')} />
         <Button title="Cerrar sesión" onPress={signOut} />
       </View>
-      <GoogleMapScreen />
-      <userLocationContext.Provider value={{ location, setLocation }}>
-        <Text>{text}</Text>
-      </userLocationContext.Provider>
     </View >
   );
 }
