+++ package.json
@@ -15,13 +15,13 @@
     "preset": "jest-expo"
   },
   "dependencies": {
-    "@expo/vector-icons": "^14.0.2",
+    "@expo/vector-icons": "^14.0.3",
     "@react-native-google-signin/google-signin": "^12.2.1",
     "@react-navigation/native": "^6.0.2",
     "expo": "~51.0.28",
     "expo-apple-authentication": "~6.4.2",
     "expo-constants": "~16.0.2",
-    "expo-font": "~12.0.9",
+    "expo-font": "~12.0.10",
     "expo-linking": "~6.3.1",
     "expo-router": "~3.5.23",
     "expo-secure-store": "~13.0.2",
@@ -44,13 +44,13 @@
     "expo-location": "~17.0.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0",
+    "@babel/core": "^7.24.0",
     "@types/jest": "^29.5.12",
-    "@types/react": "~18.2.45",
+    "@types/react": "~18.2.79",
     "@types/react-native": "^0.72.8",
     "@types/react-test-renderer": "^18.0.7",
-    "jest": "^29.2.1",
-    "jest-expo": "~51.0.3",
+    "jest": "^29.4.0",
+    "jest-expo": "^51.0.0",
     "react-test-renderer": "18.2.0",
     "typescript": "~5.3.3"
   },
